backend/requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
pymongo==3.11.4
flask-cors
python-dotenv
